// src/apps/scanner/grpc/proto-types/tracked_repo.ts
export interface TrackRepoRequest {
  id: string;
  repo: string;
  lastSeenTag: string;
}

export interface TrackRepoResponse {
  id: string;
  repo: string;
  lastSeenTag: string;
}

export interface UntrackRepoRequest {
  id: string;
}

export interface UntrackRepoResponse {
  message: string;
}

function createBaseTrackRepoRequest(): TrackRepoRequest {
  return { id: '', repo: '', lastSeenTag: '' };
}

export const TrackRepoRequest: MessageFns<TrackRepoRequest> = {
  encode(message: TrackRepoRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== '') {
      writer.uint32(10).string(message.id);
    }
    if (message.repo !== '') {
      writer.uint32(18).string(message.repo);
    }
    if (message.lastSeenTag !== '') {
      writer.uint32(26).string(message.lastSeenTag);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TrackRepoRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTrackRepoRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.repo = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.lastSeenTag = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TrackRepoRequest {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : '',
      repo: isSet(object.repo) ? globalThis.String(object.repo) : '',
      lastSeenTag: isSet(object.lastSeenTag)
        ? globalThis.String(object.lastSeenTag)
        : isSet(object.last_seen_tag)
          ? globalThis.String(object.last_seen_tag)
          : '',
    };
  },

  toJSON(message: TrackRepoRequest): unknown {
    const obj: any = {};
    if (message.id !== '') {
      obj.id = message.id;
    }
    if (message.repo !== '') {
      obj.repo = message.repo;
    }
    if (message.lastSeenTag !== '') {
      obj.lastSeenTag = message.lastSeenTag;
    }
    return obj;
  },

  create(base?: DeepPartial<TrackRepoRequest>): TrackRepoRequest {
    return TrackRepoRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TrackRepoRequest>): TrackRepoRequest {
    const message = createBaseTrackRepoRequest();
    message.id = object.id ?? '';
    message.repo = object.repo ?? '';
    message.lastSeenTag = object.lastSeenTag ?? '';
    return message;
  },
};

function createBaseTrackRepoResponse(): TrackRepoResponse {
  return { id: '', repo: '', lastSeenTag: '' };
}

export const TrackRepoResponse: MessageFns<TrackRepoResponse> = {
  encode(message: TrackRepoResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== '') {
      writer.uint32(10).string(message.id);
    }
    if (message.repo !== '') {
      writer.uint32(18).string(message.repo);
    }
    if (message.lastSeenTag !== '') {
      writer.uint32(26).string(message.lastSeenTag);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TrackRepoResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTrackRepoResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.repo = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.lastSeenTag = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TrackRepoResponse {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : '',
      repo: isSet(object.repo) ? globalThis.String(object.repo) : '',
      lastSeenTag: isSet(object.lastSeenTag)
        ? globalThis.String(object.lastSeenTag)
        : isSet(object.last_seen_tag)
          ? globalThis.String(object.last_seen_tag)
          : '',
    };
  },

  toJSON(message: TrackRepoResponse): unknown {
    const obj: any = {};
    if (message.id !== '') {
      obj.id = message.id;
    }
    if (message.repo !== '') {
      obj.repo = message.repo;
    }
    if (message.lastSeenTag !== '') {
      obj.lastSeenTag = message.lastSeenTag;
    }
    return obj;
  },

  create(base?: DeepPartial<TrackRepoResponse>): TrackRepoResponse {
    return TrackRepoResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TrackRepoResponse>): TrackRepoResponse {
    const message = createBaseTrackRepoResponse();
    message.id = object.id ?? '';
    message.repo = object.repo ?? '';
    message.lastSeenTag = object.lastSeenTag ?? '';
    return message;
  },
};

function createBaseUntrackRepoRequest(): UntrackRepoRequest {
  return { id: '' };
}

export const UntrackRepoRequest: MessageFns<UntrackRepoRequest> = {
  encode(message: UntrackRepoRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== '') {
      writer.uint32(10).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UntrackRepoRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUntrackRepoRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UntrackRepoRequest {
    return { id: isSet(object.id) ? globalThis.String(object.id) : '' };
  },

  toJSON(message: UntrackRepoRequest): unknown {
    const obj: any = {};
    if (message.id !== '') {
      obj.id = message.id;
    }
    return obj;
  },

  create(base?: DeepPartial<UntrackRepoRequest>): UntrackRepoRequest {
    return UntrackRepoRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<UntrackRepoRequest>): UntrackRepoRequest {
    const message = createBaseUntrackRepoRequest();
    message.id = object.id ?? '';
    return message;
  },
};

function createBaseUntrackRepoResponse(): UntrackRepoResponse {
  return { message: '' };
}

export const UntrackRepoResponse: MessageFns<UntrackRepoResponse> = {
  encode(message: UntrackRepoResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.message !== '') {
      writer.uint32(10).string(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UntrackRepoResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUntrackRepoResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.message = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UntrackRepoResponse {
    return { message: isSet(object.message) ? globalThis.String(object.message) : '' };
  },

  toJSON(message: UntrackRepoResponse): unknown {
    const obj: any = {};
    if (message.message !== '') {
      obj.message = message.message;
    }
    return obj;
  },

  create(base?: DeepPartial<UntrackRepoResponse>): UntrackRepoResponse {
    return UntrackRepoResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<UntrackRepoResponse>): UntrackRepoResponse {
    const message = createBaseUntrackRepoResponse();
    message.message = object.message ?? '';
    return message;
  },
};

export type TrackedRepoServiceService = typeof TrackedRepoServiceService;
export const TrackedRepoServiceService = {
  trackRepo: {
    path: '/trackedrepo.v1.TrackedRepoService/TrackRepo' as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: TrackRepoRequest): Buffer => Buffer.from(TrackRepoRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): TrackRepoRequest => TrackRepoRequest.decode(value),
    responseSerialize: (value: TrackRepoResponse): Buffer => Buffer.from(TrackRepoResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): TrackRepoResponse => TrackRepoResponse.decode(value),
  },
  untrackRepo: {
    path: '/trackedrepo.v1.TrackedRepoService/UntrackRepo' as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: UntrackRepoRequest): Buffer => Buffer.from(UntrackRepoRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): UntrackRepoRequest => UntrackRepoRequest.decode(value),
    responseSerialize: (value: UntrackRepoResponse): Buffer => Buffer.from(UntrackRepoResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): UntrackRepoResponse => UntrackRepoResponse.decode(value),
  },
} as const;

export interface TrackedRepoServiceServer extends UntypedServiceImplementation {
  trackRepo: handleUnaryCall<TrackRepoRequest, TrackRepoResponse>;
  untrackRepo: handleUnaryCall<UntrackRepoRequest, UntrackRepoResponse>;
}
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}